requirements.txt
Flask==3.0.0
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
requests==2.31.0
weasyprint==59.0
docxtpl==0.16.7
python-dotenv==1.0.0
redis==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Hashids==0.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
marshmallow-sqlalchemy==1.4.2
tenacity==8.2.3
